// practica2/types.ts
export type JobList = {
    data: Job[];
}

export type Job = {
    slug:         string;
    company_name: string;
    title:        string;
    description:  string;
    remote:       boolean;
    url:          string;
    tags:         string[];
    job_types:    string[];
    location:     string;
    created_at:   number;
}